improved application design
