+++ templates/show_data.html
@@ -86,7 +86,7 @@
     <div class="p-4 m-3 border rounded shadow">
         <h1>Service Data</h1>
         <div class="table-responsive">
-            <table id="example2" class="display nowrap table table-hover table-striped table-bordered table-sm"
+            <table id="example2" class="display table table-hover table-striped table-bordered table-sm"
                 style="width:100%">
                 <thead>
                     <tr>
@@ -113,7 +113,7 @@
                     <tr>
                         <td>{{item.id}}</td>
                         <td>{{ 'Pass' if item.status else 'Fail' }}</td>
-                        <td>{{item.message}}</td>
+                        <td class="wrap-text">{{item.message}}</td>
                         <td>{{item.ip_address}}</td>
                         <td>{{item.service_type}}</td>
                         <td>{{item.model_name}}</td>
